Changed the Logger trait in to an enum.
