+++ bld_commands/src/worker/command.rs
@@ -7,7 +7,7 @@ use awc::Client;
 use bld_config::BldConfig;
 use bld_core::database::{new_connection_pool, pipeline_runs};
 use bld_core::execution::Execution;
-use bld_core::logger::FileLogger;
+use bld_core::logger::Logger;
 use bld_core::proxies::ServerPipelineProxy;
 use bld_runner::RunnerBuilder;
 use bld_supervisor::base::WorkerMessages;
@@ -79,7 +79,7 @@ impl BldCommand for WorkerCommand {
         let pipeline_run = pipeline_runs::select_by_id(&conn, &run_id)?;
         let start_date_time = pipeline_run.start_date_time;
         let proxy = Arc::new(ServerPipelineProxy::new(cfg.clone(), pool.clone()));
-        let logger = FileLogger::atom(cfg.clone(), &run_id)?;
+        let logger = Logger::file_atom(cfg.clone(), &run_id)?;
         let exec = Execution::pipeline_atom(pool, &run_id);
         let (worker_tx, worker_rx) = channel(4096);
         let worker_tx = Arc::new(Some(worker_tx));
