--- bld_core/src/logger/base.rs
@@ -1,6 +1,0 @@
-pub trait Logger {
-    fn dump(&mut self, text: &str);
-    fn dumpln(&mut self, text: &str);
-    fn info(&mut self, text: &str);
-    fn error(&mut self, text: &str);
-}
