+++ bld_runner/src/sync/runner.rs
@@ -16,7 +16,7 @@ use tokio::sync::mpsc::Sender;
 
 type RecursiveFuture = Pin<Box<dyn Future<Output = anyhow::Result<()>>>>;
 type AtomicExec = Arc<Mutex<Execution>>;
-type AtomicLog = Arc<Mutex<dyn Logger>>;
+type AtomicLog = Arc<Mutex<Logger>>;
 type AtomicVars = Arc<HashMap<String, String>>;
 type AtomicProxy = Arc<dyn PipelineFileSystemProxy>;
